render_backend/routes/sync.py
"""
Sync routes for Render backend
"""
from fastapi import APIRouter, Depends, HTTPException, Header
from sqlalchemy.ext.asyncio import AsyncSession
from sqlalchemy.future import select
from sqlalchemy import update
from typing import List, Dict, Any, Optional
from datetime import datetime
from jose import jwt, JWTError
import os
import logging

from ..db.models import User, Notebook, SyncLog

logger = logging.getLogger(__name__)

router = APIRouter()

# Auth0 configuration
AUTH0_DOMAIN = os.getenv("AUTH0_DOMAIN")
AUTH0_AUDIENCE = os.getenv("AUTH0_AUDIENCE")
ALGORITHM = "RS256"


async def get_db():
    """Dependency to get database session"""
    # TODO: Implement async database session
    # This is a placeholder - actual implementation depends on your DB setup
    pass


def verify_token(authorization: Optional[str] = Header(None)) -> Dict[str, Any]:
    """
    Verify Auth0 JWT token
    
    Args:
        authorization: Authorization header with Bearer token
        
    Returns:
        Decoded token payload
        
    Raises:
        HTTPException: If token is invalid
    """
    if not authorization or not authorization.startswith("Bearer "):
        raise HTTPException(status_code=401, detail="Missing or invalid authorization header")
    
    token = authorization.split(" ")[1]
    
    try:
        # TODO: Implement proper JWT verification with Auth0 public key
        # For now, decode without verification (DEV ONLY!)
        payload = jwt.decode(
            token,
            options={"verify_signature": False},  # WARNING: DEV ONLY!
            audience=AUTH0_AUDIENCE,
        )
        return payload
    
    except JWTError as e:
        logger.error(f"JWT verification failed: {e}")
        raise HTTPException(status_code=401, detail="Invalid token")


@router.post("/sync")
async def sync_push(
    payload: Dict[str, Any],
    token_payload: Dict[str, Any] = Depends(verify_token),
    # db: AsyncSession = Depends(get_db)
):
    """
    Push notebooks to cloud storage
    
    Request body:
    {
        "user_id": "auth0|123456",
        "notebooks": [...],
        "last_sync": "2024-01-01T00:00:00Z"
    }
    
    Response:
    {
        "status": "synced",
        "synced_count": 5,
        "conflicts": []
    }
    """
    try:
        user_id = payload.get("user_id")
        notebooks = payload.get("notebooks", [])
        last_sync = payload.get("last_sync")
        
        # Verify user_id matches token
        token_user_id = token_payload.get("sub")
        if user_id != token_user_id:
            raise HTTPException(status_code=403, detail="User ID mismatch")
        
        # TODO: Implement actual database operations
        # For now, mock response
        logger.info(f"Syncing {len(notebooks)} notebooks for user {user_id}")
        
        synced_count = len(notebooks)
        conflicts = []
        
        # Mock: Check for conflicts (in real implementation, compare timestamps)
        # for notebook in notebooks:
        #     existing = await db.query(Notebook).filter_by(
        #         id=notebook["id"],
        #         user_id=user_id
        #     ).first()
        #     
        #     if existing and existing.updated_at > last_sync:
        #         conflicts.append({
        #             "notebook_id": notebook["id"],
        #             "server_version": existing.updated_at.isoformat()
        #         })
        
        return {
            "status": "synced",
            "synced_count": synced_count,
            "conflicts": conflicts
        }
    
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Sync push error: {e}", exc_info=True)
        raise HTTPException(status_code=500, detail="Sync failed")


@router.get("/sync")
async def sync_pull(
    user_id: str,
    token_payload: Dict[str, Any] = Depends(verify_token),
    # db: AsyncSession = Depends(get_db)
):
    """
    Pull notebooks from cloud storage
    
    Query params:
    - user_id: Auth0 user ID
    
    Response:
    {
        "notebooks": [...],
        "last_sync": "2024-01-01T00:00:00Z"
    }
    """
    try:
        # Verify user_id matches token
        token_user_id = token_payload.get("sub")
        if user_id != token_user_id:
            raise HTTPException(status_code=403, detail="User ID mismatch")
        
        # TODO: Implement actual database query
        # notebooks = await db.query(Notebook).filter_by(user_id=user_id).all()
        
        logger.info(f"Pulling notebooks for user {user_id}")
        
        # Mock response
        notebooks = []
        last_sync = datetime.utcnow().isoformat()
        
        return {
            "notebooks": notebooks,
            "last_sync": last_sync
        }
    
    except HTTPException:
        raise
    except Exception as e:
        logger.error(f"Sync pull error: {e}", exc_info=True)
        raise HTTPException(status_code=500, detail="Sync failed")


@router.get("/health")
async def health_check():
    """Health check endpoint"""
    return {
        "status": "healthy",
        "service": "clarity-sync",
        "version": "1.0.0"
    }
